requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.11
cryptography==41.0.7
python-dotenv==1.0.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
openpyxl==3.1.2
PyJWT==2.8.0
requests==2.31.0
boto3==1.34.0

